MGMT-15870: Only allow reinstallation so many times before pausing installation
As part of the ongoing effort to improve the error messaging when an installation fails, we would like to pause the installation if it has exceeded a
given threshold (clusterDeployment.Spec.InstallAttemptsLimit)

The counter will be maintained by updating the count that we find in ClusterDeployment.Status.InstallRestarts, incrementing this by one on each restart and
then storing this in AgentClusterInstall.Status.InstallRestarts.

Hive then consumes this, updates the count in ClusterDeployment.Status.InstallRestarts where we may then use this to determine whether or noth the
threshold has been exceeded.