MGMT-16944: Change parent of ClusterDeployment to Assisted Service Deployment
Presently we are expereincing an unusual issue for users who wish to remove the AgentServiceConfig and uninstall the Assisted Operator.
The incorrect parenting of the ClusterDeployment seems to prevent proper cleanup of the ClusterDeployment and AgentClusterInstall, which prevents
removal of the namespace.

This PR aims to address that by changing the parent of the ClusterDeployment to be the Assisted Service Deployment instead.